Adding `MultiSendCallOnly` during setup (#26)
## TLDR
- Added immediate delegate allowance for `MultiSendCallOnly` at the time
of guard setup.

## LLM Description
This pull request introduces a new constant, `MULTISEND_CALL_ONLY`, to
the Guardrail app and integrates it into the transaction logic for
activating guardrails. The changes primarily focus on enhancing the
functionality of the `App` component by adding a new transaction when
guardrails are activated.

### New Constant Addition:
*
[`guardrail-app/src/constants.ts`](diffhunk://#diff-c90788347f588244720a3dd7c1d303bef9d97c4f4ed76439d79c49288b7a0f3eR5):
Added `MULTISEND_CALL_ONLY` as a new constant, representing an address
used in the Guardrail app.

### Updates to the `App` Component:
*
[`guardrail-app/src/App.tsx`](diffhunk://#diff-fd339760efc333a5e935af2f83e321b0e7cdf48203ed1fe7b4620e7aab49ef64L22-R22):
Imported the `MULTISEND_CALL_ONLY` constant for use in the file.
*
[`guardrail-app/src/App.tsx`](diffhunk://#diff-fd339760efc333a5e935af2f83e321b0e7cdf48203ed1fe7b4620e7aab49ef64R222-R232):
Updated the `App` component to include a new transaction,
`immediateMultiSendCallOnlyAllowance`, in the transaction list when
activating guardrails. This transaction encodes a call to the
`immediateDelegateAllowance` function with the `MULTISEND_CALL_ONLY`
address.