Experimental Alert (#27)
This pull request adds a new warning alert to the top of the application
to inform users that the demo is an experimental beta release and that
the code is not audited. This improves user awareness about the risks of
using the application.

User-facing improvements:

* Added a warning `Alert` component at the top of the main page in
`App.tsx` to notify users that the demo is experimental and unaudited.